+++ lib/types/RuntimeInterfaces.ts
@@ -1,14 +1,11 @@
 
+import type { StateObject } from "./StateObjectTypes"
+
 export type EffectType = {
    callback: () => void | (() => void),
    dependencies: Array<unknown>|null,
    cleanup: (() => void)|null,
    lastDeps: unknown[]|null
-}
-
-export interface EventObject {
-   beforeload: ((payload: EventPayload) => void)[],
-   load: ((payload: EventPayload) => void)[],
 }
 
 export interface EventPayload {
@@ -18,9 +15,27 @@ export interface EventPayload {
    data?: string
 }
 
+export interface PopStatePayload {
+   route: string,
+   state: StateObject | null,
+   nativeEvent: PopStateEvent,
+   defaultPrevented: boolean,
+   preventDefault: () => void,
+}
+
+export interface EventPayloadMap {
+   beforeload: EventPayload,
+   load: EventPayload,
+   popstate: PopStatePayload,
+}
+
 export interface RuntimeConfig {
    preserveUpdatedHtmlState: boolean,
    waitForStyles: boolean
+}
+
+export type EventObject = {
+   [K in keyof EventPayloadMap]: Array<(payload: EventPayloadMap[K]) => void>
 }
 
 export type CurrentRoutesObject = Record<string, {